Forbidden gif for David

+++ index.js
@@ -145,7 +145,11 @@ client.on("message", async (message) => {
     return;
   }
 
-  if (message.content.includes(".gif") || message.content.includes("-gif"))
+  if (
+    (message.content.includes(".gif") || message.content.includes("-gif")) &&
+    message.author.id ===
+      client.users.cache.find((u) => u.tag === "3T33TT#6772").id
+  )
     return message.delete();
 
   // Img Suppr
